fixed a subtle bug in consensus
When validators vote for a block, they use the last finalized block as a justification for their vote,
however that block may not have been finalized already on all validators because of delays in blocks and vote propagation,
originally validators required the justification to be a finalized block, and that led to a situation where validators that
didn't accumulate all the nessesary vote to finalize a block that was used as a justification by the voting validator to always
reject the vote and in turn would stop finalizing and confirming blocks getting stuck on that one rejected vote that would be needed
to hit 2/3 voting stake.

This change accepts justifications not only for finalized blocks but also for any descendant of the target of the vote. Now that this
rule changed, it is SECURITY CRITICAL to implement a detection mechanism for the second slashable condition in CBC Casper: Surround vote,
which means that a validator could never use an older justofication that a previously used justification when voting for a block.

New tasks will be created and assigned as part of a more comprehensive slashing & rewards work item.