Explicit block reply mechanism
Noticed that after few k confirmed/finalized blocks the consensus gets stuck and stops confirming blocks because of a missing block that was lost by the gossip somewhere.
This change adds an explicit request to replay a given block from the consensus forktree (volatile blocks) if the block has been missing and not recovered automatically by the p2p repair mechanism for longer than half an epoch.