Transaction Ordering in Mempool and Block Producer
This change introduces a deterministic transaction ordering by their transaction hash, except when transactions
are coming from the same payer, then they are ordered by the nonce value. Blocks that do not adhere to this tx
ordering algorithm are rejected by all validators.

This change SIGNIFICANTLY hinders MEV and protects users of the chain from malicious bots.

Another change that was introduced is the ability to send a batch of transactions with one RPC call, this is for
scenarios like:

 - Payroll, Mass Transfers, Group Airdrops, etc.
 - Contract Upload, contract bytecode is split into small chunks that fit into transactions and then transmitted
   in pieces. This allows packing many pieces within one block instead of having each tx to confirm before its
   nonce is incremented.

At this point please igore the quality and structure of the JS API, it's going to be rewritten from scratch anyway.